+++ Tools/imgCIF_Creator/imgCIF_Creator/information_extractors/__init__.py
@@ -1,0 +1,1 @@
+"Functionalities to extract information from different input formats."
