+++ Tools/imgCIF_Creator/imgCIF_Creator/output_assembler/__init__.py
@@ -1,0 +1,2 @@
+"""Assembel imgCIF files from different parts.
+"""
